--- README.md
@@ -1,2 +1,0 @@
-# dm-lab1
-Lab No. 1 for Discrete math course(UCU).
